+++ Core/Inc/porting.h
@@ -14,16 +14,11 @@
 #include <math.h>
 #include <limits.h>
 
-#define IEXTFLASH_ATTR __attribute__((section (".extflash_text")))
-#define DEXTFLASH_ATTR __attribute__((section (".extflash_data")))
+#define IEXTFLASH_ATTR
+#define DEXTFLASH_ATTR
 
-#ifdef PORTING_USE_EXTFLASH
-#   define IRAM_ATTR __attribute__((section (".ram_text")))
-#   define DRAM_ATTR __attribute__((section (".ram_data")))
-#else
-#   define IRAM_ATTR
-#   define DRAM_ATTR
-#endif
+#define IRAM_ATTR
+#define DRAM_ATTR
 
 #ifndef DEBUG_RG_ALLOC
