+++ Core/Src/porting/gb/main_gb.c
@@ -51,7 +51,7 @@ static bool saveSRAM = false;
 static int  saveSRAM_Timer = 0;
 
 // 3 pages
-uint8_t state_save_buffer[192 * 1024] __attribute__((section (".emulator_data")));
+uint8_t state_save_buffer[192 * 1024];
 
 
 // --- MAIN
@@ -484,23 +484,6 @@ void pcm_submit() {
             audiobuffer_dma[i + offset] = pcm.buf[i] >> shift;
         }
     }
-}
-
-bool odroid_netplay_quick_start(void)
-{
-    return true;
-}
-
-// TODO: Move to own file
-void odroid_audio_mute(bool mute)
-{
-    if (mute) {
-        for (int i = 0; i < sizeof(audiobuffer_dma) / sizeof(audiobuffer_dma[0]); i++) {
-            audiobuffer_dma[i] = 0;
-        }
-    }
-
-    audio_mute = mute;
 }
