+++ Core/Src/porting/nes/main_nes.c
@@ -37,7 +37,7 @@ static bool autoload = false;
 
 
 // if i counted correctly this should max be 23077
-uint8_t nes_save_buffer[24000] __attribute__((section (".emulator_data")));
+uint8_t nes_save_buffer[24000];
 
 // TODO: Expose properly
 extern int nes_state_save(uint8_t *flash_ptr, size_t size);
