+++ Core/Src/retro-go/rg_emulators.c
@@ -2,6 +2,7 @@
 #include <string.h>
 #include <assert.h>
 
+#include "gw_linker.h"
 #include "rg_emulators.h"
 // #include "rg_favorites.h"
 #include "bitmaps.h"
@@ -373,8 +374,14 @@ void emulator_start(retro_emulator_file_t *file, bool load_state)
     retro_emulator_t *emu = file->emulator;
     // TODO: Make this cleaner
     if(strcmp(emu->system_name, "Nintendo Gameboy") == 0) {
+        memcpy(&__RAM_EMU_START__, &_OVERLAY_GB_LOAD_START, (size_t)&_OVERLAY_GB_SIZE);
+        memset(&_OVERLAY_GB_BSS_START, 0x0, (size_t)&_OVERLAY_GB_BSS_SIZE);
+        SCB_CleanDCache_by_Addr((uint32_t *)&__RAM_EMU_START__, (size_t)&_OVERLAY_GB_SIZE);
         app_main_gb(load_state);
     } else if(strcmp(emu->system_name, "Nintendo Entertainment System") == 0) {
+        memcpy(&__RAM_EMU_START__, &_OVERLAY_NES_LOAD_START, (size_t)&_OVERLAY_NES_SIZE);
+        memset(&_OVERLAY_NES_BSS_START, 0x0, (size_t)&_OVERLAY_NES_BSS_SIZE);
+        SCB_CleanDCache_by_Addr((uint32_t *)&__RAM_EMU_START__, (size_t)&_OVERLAY_NES_SIZE);
         app_main_nes(load_state);
     }
